nss: ensure DomesticPolicy ciphers can be applied
NSS_SetDomesticPolicy() is not implemented as an atomic operation
where either all ciphers work or none, so it's possible that that
only some get applied during a call.  If so, the returnvalue will
be distinct from SECSuccess to indicate failure.  Thus, we must
inspect the returnvalue to ensure that all the expected ciphers
are applied.